backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database Drivers
neo4j==5.17.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis and Task Queue
redis==7.2.1
celery==5.3.6
flower==2.0.1

# OpenAI and AI/ML
openai==1.12.0
tiktoken==0.5.2
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# Data Processing
pandas==2.2.0
numpy==1.26.3
feedparser==6.0.10
beautifulsoup4==4.12.3
newspaper3k==0.2.8
requests==2.31.0
aiohttp==3.9.1

# NLP and Text Processing
spacy==3.7.4
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl

# Data Validation and Settings
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# API Documentation
jsonschema==4.21.1
